+++ Cabal-tests/tests/UnitTests/Distribution/Simple/Glob.hs
@@ -1,22 +1,29 @@
 {-# LANGUAGE LambdaCase #-}
-module UnitTests.Distribution.Simple.Glob
-    ( tests
-    ) where
 
-import Control.Monad
-import Data.Foldable (for_)
-import Data.Function (on)
-import Data.List (sort)
-import Data.Maybe (mapMaybe)
-import Distribution.Simple.Glob
-import qualified Distribution.Verbosity as Verbosity
+module UnitTests.Distribution.Simple.Glob
+  ( tests
+  ) where
+
+import Distribution.Compat.Prelude hiding (last)
+import Prelude ()
+
 import Distribution.CabalSpecVersion
+import Distribution.Parsec
+import Distribution.Pretty
+import Distribution.Simple.FileMonitor.Types
+import Distribution.Simple.Glob
+import Distribution.Simple.Glob.Internal
+import Distribution.Utils.Structured (structureHash)
+import Distribution.Verbosity
+import Distribution.Verbosity qualified as Verbosity
+import GHC.Fingerprint
 import System.Directory (createDirectoryIfMissing)
-import System.FilePath ((</>), splitFileName, normalise)
+import System.FilePath (normalise, splitFileName, (</>))
 import System.IO.Temp (withSystemTempDirectory)
+import Test.QuickCheck.Instances.Cabal ()
 import Test.Tasty
 import Test.Tasty.HUnit
-import Distribution.Verbosity
+import Test.Tasty.QuickCheck
 
 sampleFileNames :: [FilePath]
 sampleFileNames =
@@ -51,6 +58,134 @@ makeSampleFiles dir = for_ sampleFileNames $ \filename -> do
   createDirectoryIfMissing True (dir </> dir')
   writeFile (dir </> dir' </> name) $ "This is " ++ filename
 
+prop_roundtrip_printparse :: RootedGlob -> Property
+prop_roundtrip_printparse pathglob =
+  counterexample (prettyShow pathglob) $
+    eitherParsec (prettyShow pathglob) === Right pathglob
+
+-- first run, where we don't even call updateMonitor
+testParseCases :: Assertion
+testParseCases = do
+  RootedGlob (FilePathRoot "/") GlobDirTrailing <- testparse "/"
+  RootedGlob FilePathHomeDir GlobDirTrailing <- testparse "~/"
+
+  RootedGlob (FilePathRoot "A:\\") GlobDirTrailing <- testparse "A:/"
+  RootedGlob (FilePathRoot "Z:\\") GlobDirTrailing <- testparse "z:/"
+  RootedGlob (FilePathRoot "C:\\") GlobDirTrailing <- testparse "C:\\"
+  RootedGlob FilePathRelative (GlobFile [Literal "_:"]) <- testparse "_:"
+
+  RootedGlob
+    FilePathRelative
+    (GlobDirRecursive [WildCard]) <-
+    testparse "**/*"
+  RootedGlob
+    FilePathRelative
+    (GlobDir [Literal "foo"] (GlobDirRecursive [WildCard])) <-
+    testparse "foo/**/*"
+  RootedGlob
+    FilePathRelative
+    (GlobDir [Literal "foo"] (GlobDirRecursive [WildCard, Literal ".txt"])) <-
+    testparse "foo/**/*.txt"
+
+  RootedGlob
+    FilePathRelative
+    (GlobFile [Literal "."]) <-
+    testparse "."
+
+  RootedGlob
+    FilePathRelative
+    (GlobFile [Literal "~"]) <-
+    testparse "~"
+
+  RootedGlob
+    FilePathRelative
+    (GlobDir [Literal "."] GlobDirTrailing) <-
+    testparse "./"
+
+  RootedGlob
+    FilePathRelative
+    (GlobFile [Literal "foo"]) <-
+    testparse "foo"
+
+  RootedGlob
+    FilePathRelative
+    ( GlobDir
+        [Literal "foo"]
+        (GlobFile [Literal "bar"])
+      ) <-
+    testparse "foo/bar"
+
+  RootedGlob
+    FilePathRelative
+    ( GlobDir
+        [Literal "foo"]
+        (GlobDir [Literal "bar"] GlobDirTrailing)
+      ) <-
+    testparse "foo/bar/"
+
+  RootedGlob
+    (FilePathRoot "/")
+    ( GlobDir
+        [Literal "foo"]
+        (GlobDir [Literal "bar"] GlobDirTrailing)
+      ) <-
+    testparse "/foo/bar/"
+
+  RootedGlob
+    (FilePathRoot "C:\\")
+    ( GlobDir
+        [Literal "foo"]
+        (GlobDir [Literal "bar"] GlobDirTrailing)
+      ) <-
+    testparse "C:\\foo\\bar\\"
+
+  RootedGlob
+    FilePathRelative
+    (GlobFile [WildCard]) <-
+    testparse "*"
+
+  RootedGlob
+    FilePathRelative
+    (GlobFile [WildCard, WildCard]) <-
+    testparse "**" -- not helpful, but valid (?)
+  RootedGlob
+    FilePathRelative
+    (GlobFile [WildCard, Literal "foo", WildCard]) <-
+    testparse "*foo*"
+
+  RootedGlob
+    FilePathRelative
+    (GlobFile [Literal "foo", WildCard, Literal "bar"]) <-
+    testparse "foo*bar"
+
+  RootedGlob
+    FilePathRelative
+    (GlobFile [Union [[WildCard], [Literal "foo"]]]) <-
+    testparse "{*,foo}"
+
+  parseFail "{"
+  parseFail "}"
+  parseFail ","
+  parseFail "{"
+  parseFail "{{}"
+  parseFail "{}"
+  parseFail "{,}"
+  parseFail "{foo,}"
+  parseFail "{,foo}"
+
+  return ()
+
+testparse :: String -> IO RootedGlob
+testparse s =
+  case eitherParsec s of
+    Right p -> return p
+    Left err -> throwIO $ HUnitFailure Nothing ("expected parse of: " ++ s ++ " -- " ++ err)
+
+parseFail :: String -> Assertion
+parseFail s =
+  case eitherParsec s :: Either String RootedGlob of
+    Right p -> throwIO $ HUnitFailure Nothing ("expected no parse of: " ++ s ++ " -- " ++ show p)
+    Left _ -> return ()
 compatibilityTests :: CabalSpecVersion -> [TestTree]
 compatibilityTests version =
   [ testCase "literal match" $
@@ -109,19 +244,23 @@ testMatchesVersion version pat expected = do
           -- check can't identify that kind of match.
           expected' = filter (\case GlobMatchesDirectory _ -> False; _ -> True) expected
       unless (sort expected' == sort actual) $
-        assertFailure $ "Unexpected result (pure matcher): " ++ show actual ++ "\nExpected: " ++ show expected
+        assertFailure $
+          "Unexpected result (pure matcher): " ++ show actual ++ "\nExpected: " ++ show expected
     checkIO globPat =
       withSystemTempDirectory "globstar-sample" $ \tmpdir -> do
         makeSampleFiles tmpdir
         actual <- runDirFileGlob verbosity (Just version) tmpdir globPat
         unless (isEqual actual expected) $
-          assertFailure $ "Unexpected result (impure matcher): " ++ show actual ++ "\nExpected: " ++ show expected
+          assertFailure $
+            "Unexpected result (impure matcher): " ++ show actual ++ "\nExpected: " ++ show expected
 
 testFailParseVersion :: CabalSpecVersion -> FilePath -> GlobSyntaxError -> Assertion
 testFailParseVersion version pat expected =
   case parseFileGlob version pat of
-    Left err -> unless (expected == err) $
-      assertFailure $ "Unexpected error: " ++ show err
+    Left err ->
+      unless (expected == err) $
+        assertFailure $
+          "Unexpected error: " ++ show err
     Right _ -> assertFailure "Unexpected success in parsing."
 
 globstarTests :: [TestTree]
@@ -157,9 +296,25 @@ multiDotTests =
   where
     testMatches = testMatchesVersion CabalSpecV2_4
 
+printParseTests :: [TestTree]
+printParseTests =
+  [ testProperty "print/parse roundtrip" prop_roundtrip_printparse
+  , testCase "parse examples" testParseCases
+  , testGroup
+      "Structured hashes"
+      [ testCase "GlobPiece" $ structureHash (Proxy :: Proxy GlobPiece) @?= Fingerprint 0xd5e5361866a30ea2 0x31fbfe7b58864782
+      , testCase "Glob" $ structureHash (Proxy :: Proxy Glob) @?= Fingerprint 0x3a5af41e8194eaa3 0xd8e461fdfdb0e07b
+      , testCase "FilePathRoot" $ structureHash (Proxy :: Proxy FilePathRoot) @?= Fingerprint 0x713373d51426ec64 0xda7376a38ecee5a5
+      , testCase "RootedGlob" $ structureHash (Proxy :: Proxy RootedGlob) @?= Fingerprint 0x0031d198379cd1bf 0x7246ab9b6c6e0e7d
+      ]
+  ]
+
 tests :: [TestTree]
 tests =
-  [ testGroup "pre-2.4 compatibility" $
+  [ testGroup
+      "print/parse"
+      printParseTests
+  , testGroup "pre-2.4 compatibility" $
       compatibilityTests CabalSpecV2_2
   , testGroup "post-2.4 compatibility" $
       compatibilityTests CabalSpecV2_4
@@ -168,3 +323,5 @@ tests =
       testFailParseVersion CabalSpecV1_4 "foo/*.bar" VersionDoesNotSupportGlob
   , testGroup "multi-dot globbing" multiDotTests
   ]
+
+-- TODO: tests for windows style file paths
