+++ Cabal/src/Distribution/Simple/Utils.hs
@@ -201,6 +201,10 @@ module Distribution.Simple.Utils
   , wrapText
   , wrapLine
   , stripCommonPrefix
+  , MergeResult (..)
+  , mergeBy
+  , duplicates
+  , duplicatesBy
 
     -- * FilePath stuff
   , isAbsoluteOnAnyPlatform
@@ -240,6 +244,9 @@ import Distribution.Parsec
 import Distribution.Pretty
 
 import qualified Data.ByteString.Lazy as BS
+import Data.List
+  ( groupBy
+  )
 import Data.Typeable
   ( cast
   )
@@ -2171,3 +2178,31 @@ stripCommonPrefix (x : xs) (y : ys)
   | x == y = stripCommonPrefix xs ys
   | otherwise = y : ys
 stripCommonPrefix _ ys = ys
+
+-- | Generic merging utility. For sorted input lists this is a full outer join.
+mergeBy :: forall a b. (a -> b -> Ordering) -> [a] -> [b] -> [MergeResult a b]
+mergeBy cmp = merge
+  where
+    merge :: [a] -> [b] -> [MergeResult a b]
+    merge [] ys = [OnlyInRight y | y <- ys]
+    merge xs [] = [OnlyInLeft x | x <- xs]
+    merge (x : xs) (y : ys) =
+      case x `cmp` y of
+        GT -> OnlyInRight y : merge (x : xs) ys
+        EQ -> InBoth x y : merge xs ys
+        LT -> OnlyInLeft x : merge xs (y : ys)
+
+data MergeResult a b = OnlyInLeft a | InBoth a b | OnlyInRight b
+
+duplicates :: Ord a => [a] -> [NonEmpty a]
+duplicates = duplicatesBy compare
+
+duplicatesBy :: forall a. (a -> a -> Ordering) -> [a] -> [NonEmpty a]
+duplicatesBy cmp = mapMaybe moreThanOne . groupBy eq . sortBy cmp
+  where
+    eq :: a -> a -> Bool
+    eq a b = case cmp a b of
+      EQ -> True
+      _ -> False
+    moreThanOne (x : xs@(_ : _)) = Just (x :| xs)
+    moreThanOne _ = Nothing
