+++ cabal-install/src/Distribution/Client/Utils.hs
@@ -62,13 +62,21 @@ import Data.Bits
 import qualified Data.ByteString.Lazy as BS
 import Data.List
   ( elemIndex
-  , groupBy
   )
 import Distribution.Client.Errors
 import Distribution.Compat.Environment
 import Distribution.Compat.Time (getModTime)
 import Distribution.Simple.Setup (Flag, pattern Flag, pattern NoFlag)
-import Distribution.Simple.Utils (dieWithException, findPackageDesc, noticeNoWrap, removeFileForcibly)
+import Distribution.Simple.Utils
+  ( MergeResult (..)
+  , dieWithException
+  , duplicates
+  , duplicatesBy
+  , findPackageDesc
+  , mergeBy
+  , noticeNoWrap
+  , removeFileForcibly
+  )
 import Distribution.Utils.Path
   ( CWD
   , FileOrDir (..)
@@ -118,34 +126,6 @@ import GHC.IO.Encoding.Failure
   )
 import qualified System.Directory as Dir
 import qualified System.IO.Error as IOError
-
--- | Generic merging utility. For sorted input lists this is a full outer join.
-mergeBy :: forall a b. (a -> b -> Ordering) -> [a] -> [b] -> [MergeResult a b]
-mergeBy cmp = merge
-  where
-    merge :: [a] -> [b] -> [MergeResult a b]
-    merge [] ys = [OnlyInRight y | y <- ys]
-    merge xs [] = [OnlyInLeft x | x <- xs]
-    merge (x : xs) (y : ys) =
-      case x `cmp` y of
-        GT -> OnlyInRight y : merge (x : xs) ys
-        EQ -> InBoth x y : merge xs ys
-        LT -> OnlyInLeft x : merge xs (y : ys)
-
-data MergeResult a b = OnlyInLeft a | InBoth a b | OnlyInRight b
-
-duplicates :: Ord a => [a] -> [NonEmpty a]
-duplicates = duplicatesBy compare
-
-duplicatesBy :: forall a. (a -> a -> Ordering) -> [a] -> [NonEmpty a]
-duplicatesBy cmp = mapMaybe moreThanOne . groupBy eq . sortBy cmp
-  where
-    eq :: a -> a -> Bool
-    eq a b = case cmp a b of
-      EQ -> True
-      _ -> False
-    moreThanOne (x : xs@(_ : _)) = Just (x :| xs)
-    moreThanOne _ = Nothing
 
 -- | A variant of 'withTempFile' that only gives us the file name, and while
 -- it will clean up the file afterwards, it's lenient if the file is
