+++ cabal-install/tests/LongTests.hs
@@ -3,7 +3,6 @@ module Main (main) where
 import Test.Tasty
 
 import qualified UnitTests.Distribution.Client.Described
-import qualified UnitTests.Distribution.Client.FileMonitor
 import qualified UnitTests.Distribution.Client.VCS
 import qualified UnitTests.Distribution.Solver.Modular.QuickCheck
 import UnitTests.Options
@@ -24,7 +23,5 @@ tests =
           UnitTests.Distribution.Solver.Modular.QuickCheck.tests
       , testGroup "UnitTests.Distribution.Client.VCS" $
           UnitTests.Distribution.Client.VCS.tests mtimeChange
-      , testGroup "UnitTests.Distribution.Client.FileMonitor" $
-          UnitTests.Distribution.Client.FileMonitor.tests mtimeChange
       , UnitTests.Distribution.Client.Described.tests
       ]
