+++ cabal-install/tests/UnitTests/Distribution/Client/ArbitraryInstances.hs
@@ -23,8 +23,6 @@ module UnitTests.Distribution.Client.ArbitraryInstances
 import Distribution.Client.Compat.Prelude
 import Prelude ()
 
-import Data.Char (isLetter)
-import Data.List ((\\))
 import Data.Monoid (Last (..))
 
 import Distribution.Simple.Setup
@@ -32,7 +30,6 @@ import Distribution.Types.Flag (mkFlagAssignment)
 
 import Distribution.Client.BuildReports.Types (BuildReport, InstallOutcome, Outcome, ReportLevel (..))
 import Distribution.Client.CmdInstall.ClientInstallFlags (InstallMethod)
-import Distribution.Client.Glob (FilePathRoot (..), Glob (..), GlobPiece (..), RootedGlob (..))
 import Distribution.Client.IndexUtils.ActiveRepos (ActiveRepoEntry (..), ActiveRepos (..), CombineStrategy (..))
 import Distribution.Client.IndexUtils.IndexState (RepoIndexState (..), TotalIndexState, makeTotalIndexState)
 import Distribution.Client.IndexUtils.Timestamp (Timestamp, epochTimeToTimestamp)
@@ -48,7 +45,6 @@ import Network.URI (URI (..), URIAuth (..), isUnreserved)
 import Test.QuickCheck
   ( Arbitrary (..)
   , Gen
-  , NonEmptyList (..)
   , arbitraryBoundedEnum
   , choose
   , elements
@@ -338,114 +334,3 @@ instance Arbitrary InstallOutcome where
 instance Arbitrary Outcome where
   arbitrary = genericArbitrary
   shrink = genericShrink
-
--------------------------------------------------------------------------------
--- Glob
--------------------------------------------------------------------------------
-
-instance Arbitrary RootedGlob where
-  arbitrary =
-    (RootedGlob <$> arbitrary <*> arbitrary)
-      `suchThat` validFilePathGlob
-
-  shrink (RootedGlob root pathglob) =
-    [ RootedGlob root' pathglob'
-    | (root', pathglob') <- shrink (root, pathglob)
-    , validFilePathGlob (RootedGlob root' pathglob')
-    ]
-
-validFilePathGlob :: RootedGlob -> Bool
-validFilePathGlob (RootedGlob FilePathRelative pathglob) =
-  case pathglob of
-    GlobDirTrailing -> False
-    GlobDir [Literal "~"] _ -> False
-    GlobDir [Literal (d : ":")] _
-      | isLetter d -> False
-    _ -> True
-validFilePathGlob _ = True
-
-instance Arbitrary FilePathRoot where
-  arbitrary =
-    frequency
-      [ (3, pure FilePathRelative)
-      , (1, pure (FilePathRoot unixroot))
-      , (1, FilePathRoot <$> windrive)
-      , (1, pure FilePathHomeDir)
-      ]
-    where
-      unixroot = "/"
-      windrive = do d <- choose ('A', 'Z'); return (d : ":\\")
-
-  shrink FilePathRelative = []
-  shrink (FilePathRoot _) = [FilePathRelative]
-  shrink FilePathHomeDir = [FilePathRelative]
-
-instance Arbitrary Glob where
-  arbitrary = sized $ \sz ->
-    oneof $
-      take
-        (max 1 sz)
-        [ pure GlobDirTrailing
-        , GlobFile . getGlobPieces <$> arbitrary
-        , (GlobDir . getGlobPieces <$> arbitrary)
-            <*> resize (sz `div` 2) arbitrary
-        ]
-
-  shrink GlobDirTrailing = []
-  shrink (GlobFile glob) =
-    GlobDirTrailing
-      : [GlobFile (getGlobPieces glob') | glob' <- shrink (GlobPieces glob)]
-  shrink (GlobDir glob pathglob) =
-    pathglob
-      : GlobFile glob
-      : [ GlobDir (getGlobPieces glob') pathglob'
-        | (glob', pathglob') <- shrink (GlobPieces glob, pathglob)
-        ]
-  shrink (GlobDirRecursive glob) =
-    GlobDirTrailing
-      : [GlobFile (getGlobPieces glob') | glob' <- shrink (GlobPieces glob)]
-
-newtype GlobPieces = GlobPieces {getGlobPieces :: [GlobPiece]}
-  deriving (Eq)
-
-instance Arbitrary GlobPieces where
-  arbitrary = GlobPieces . mergeLiterals <$> shortListOf1 5 arbitrary
-
-  shrink (GlobPieces glob) =
-    [ GlobPieces (mergeLiterals (getNonEmpty glob'))
-    | glob' <- shrink (NonEmpty glob)
-    ]
-
-mergeLiterals :: [GlobPiece] -> [GlobPiece]
-mergeLiterals (Literal a : Literal b : ps) = mergeLiterals (Literal (a ++ b) : ps)
-mergeLiterals (Union as : ps) = Union (map mergeLiterals as) : mergeLiterals ps
--- Two consecutive wildcards are semantically equivalent to a single one, but
--- would syntactically produce a recursive wildcard when pretty-printed, so
--- whenever we end up generating two or more consecutive wildcards, we merge
--- them together to avoid this problem.
-mergeLiterals (WildCard : WildCard : ps) = mergeLiterals (WildCard : ps)
-mergeLiterals (p : ps) = p : mergeLiterals ps
-mergeLiterals [] = []
-
-instance Arbitrary GlobPiece where
-  arbitrary = sized $ \sz ->
-    frequency
-      [ (3, Literal <$> shortListOf1 10 (elements globLiteralChars))
-      , (1, pure WildCard)
-      , (1, Union <$> resize (sz `div` 2) (shortListOf1 5 (shortListOf1 5 arbitrary)))
-      ]
-
-  shrink (Literal str) =
-    [ Literal str'
-    | str' <- shrink str
-    , not (null str')
-    , all (`elem` globLiteralChars) str'
-    ]
-  shrink WildCard = []
-  shrink (Union as) =
-    [ Union (map getGlobPieces (getNonEmpty as'))
-    | as' <- shrink (NonEmpty (map GlobPieces as))
-    ]
-
-globLiteralChars :: [Char]
-globLiteralChars = ['\0' .. '\128'] \\ "*{},/\\"
