+++ changelog.d/12065.md
@@ -1,0 +1,22 @@
+---
+synopsis: Code restructuring between Cabal and cabal-install
+packages: [Cabal, cabal-install]
+prs: 12065
+---
+
+Moved file-monitoring code from cabal-install into Cabal, and relevant tests
+from cabal-install tests to Cabal-tests.
+
+The following functionality is now available through the Cabal library:
+
+- File monitoring
+- Globs
+- Hash functions
+
+Additionally, a typeclass MonadRebuild has been introduced, which captures the
+generic behavior of the Rebuild monad; most functionality from
+cabal-install that uses the Rebuild monad has been rewritten to only depend on
+MonadRebuild, and moved to Cabal.
+
+To keep code backwards compatible, all moved / rewritten functions are
+re-exported under their old names from their previous modules.
